% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityClust.R
\name{clustered}
\alias{clustered}
\alias{clustered.densityCluster}
\title{Check whether a densityCluster object have been clustered}
\usage{
clustered(x)

\method{clustered}{densityCluster}(x)
}
\arguments{
\item{x}{A densityCluster object}
}
\value{
TRUE if \code{\link{findClusters}} have been performed, otherwise 
FALSE
}
\description{
This function checks whether \code{\link{findClusters}} has been performed on
the given object and returns a boolean depending on the outcome
}
