% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clusters}
\alias{clusters}
\alias{clusters.densityCluster}
\title{Extract cluster membership from a densityCluster object}
\usage{
clusters(x, ...)

\method{clusters}{densityCluster}(x, as.list = FALSE, halo.rm = TRUE, ...)
}
\arguments{
\item{x}{the \code{densityCluster} object to inspect.}

\item{...}{extra arguments to pass on.}

\item{as.list}{logical indicating that the output should be in list format.}

\item{halo.rm}{logical indicating that halo points should be removed.}
}
\value{
A vector or list with cluster memberships for the observations in the
  initial distance matrix
}
\description{
This function allows the user to extract the cluster membership of all the
observations in the given \code{densityCluster} object. The output can be
formatted in two ways as described below. Halo observations can optionally be
removed from the output.
}
\details{
Two formats for the output are available, with the first being a vector of
integers denoting for each point which cluster the point belongs
to. If halo observations are removed, these are set to NA. The second format
is a list with a vector for each group containing the index for the member
points in the group. If halo points are removed their indexes are
omitted. The list format corresponds to the transformation of the vector
format \code{split(1:length(clusters), clusters)}, where \code{clusters} is
the cluster information in vector format.
}

