% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dclust}
\alias{dclust}
\title{Calculate clustering attributes}
\usage{
dclust(distance, dc, gaussian = FALSE)
}
\arguments{
\item{distance}{a distance matrix.}

\item{dc}{a distance cutoff for calculating the local density. If missing it
will be estimated with \code{\link{estimateDC}(distance)}.}

\item{gaussian}{logical indicating that a Gaussian kernel should be used to
  estimate the density.}
}
\value{
A densityCluster object. See Details for a description.
}
\description{
This function takes a distance matrix and optionally a distance cutoff and
calculates the values necessary for clustering based on the algorithm
proposed by Alex Rodrigues and Alessandro Laio (see \code{CITATION}). The
actual assignment to clusters is done in a later step, based on user defined
threshold values.
}
\details{
The function calculates \eqn{\rho} and \eqn{\delta} for the
  observations in the provided distance matrix. If a distance cutoff is not
  provided this is first estimated using \code{\link{estimateDC}} with
  default values.

  The information kept in the densityCluster object is: \describe{
  \item{rho}{a vector of local density values.} \item{delta}{a vector of
  minimum distances to observations of higher density.} \item{distance}{the
  initial distance matrix.} \item{dc}{the distance cutoff used to calculate
  rho.} \item{threshold}{a named vector specifying the threshold values for
  rho and delta used for cluster detection.} \item{peaks}{a vector of indices
  specifying the cluster center for each cluster.} \item{clusters}{a vector
  of cluster affiliations for each point, with the clusters being referenced as
  indices in the peaks vector.} \item{halo}{a logical vector specifying for
  each observation if it is considered part of the halo.} } Before running
  \code{\link{findClusters}} the threshold, peaks, clusters and halo data is
  \code{NA}.
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- dclust(irisDist, gaussian=TRUE)
plot(irisClust) # Inspect clustering attributes to define thresholds

irisClust <- findClusters(irisClust, rho=2, delta=2)
plotMDS(irisClust)
split(iris[,5], irisClust$clusters)
}
\seealso{
\code{\link{estimateDC}}, \code{\link{findClusters}}
}

