% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distanceToPeak}
\alias{distanceToPeak}
\title{Calculate distance to closest observation of higher density}
\usage{
distanceToPeak(distance, rho)
}
\arguments{
\item{distance}{a distance matrix.}

\item{rho}{a vector of local density values as outputted by \code{\link{localDensity}}.}
}
\value{
A vector of distances with index matching the index in \code{rho}.
}
\description{
Calculate distance to closest observation of higher density
}

