% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimateDC}
\alias{estimateDC}
\title{Estimate the distance cutoff for a specified neighbor rate}
\usage{
estimateDC(distance, neighborRateLow = 0.01, neighborRateHigh = 0.02)
}
\arguments{
\item{distance}{a distance matrix.}

\item{neighborRateLow}{the lower bound of the neighbor rate.}

\item{neighborRateHigh}{the upper bound of the neighbor rate.}
}
\value{
A numeric value giving the estimated best distance cutoff value.
}
\description{
This function calculates a distance cutoff value for a specific distance
matrix that makes the average neighbor rate (number of points within the
distance cutoff value) fall between the provided range. The authors of the
algorithm suggests aiming for a neighbor rate between 1 and 2 percent, but
also state that the algorithm is quite robust with regards to more extreme
cases.
}
\examples{
irisDist <- dist(iris[,1:4])
estimateDC(irisDist)
}

