% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findClusters}
\alias{findClusters}
\alias{findClusters.densityCluster}
\title{Calculate clusters in a densityCluster object}
\usage{
findClusters(x, ...)

\method{findClusters}{densityCluster}(x, rho, delta, plot = FALSE, ...)
}
\arguments{
\item{x}{a \code{densityCluster} object as produced by \code{\link{dclust}}.}

\item{...}{additional parameters passed to pass on.}

\item{rho}{the threshold for local density.}

\item{delta}{the threshold for minimum distance to higher density.}

\item{plot}{logical indicating whether a decision plot should be shown after
  cluster calculation.}
}
\value{
A \code{densityCluster} object with clusters assigned to all
  observations.
}
\description{
This function uses the supplied \eqn{\rho} and \eqn{\delta} thresholds to
calculate cluster centers and assigns the rest of the points to one of these
clusters, designating these points as belonging to either the core or the
halo of the cluster. If either of the \eqn{\rho} or \eqn{\delta} thresholds
are missing the user is presented with a decision plot where they are able to
click on the plot area to set the threshold. If either of the \eqn{\rho} or
\eqn{\delta} thresholds is passed as an argument then this takes precedence
over the value set by clicking.
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- dclust(irisDist, gaussian=TRUE)
irisClust <- findClusters(irisClust, rho=2, delta=2)
}

