% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.densityCluster.parallel.R
\name{findDistValueByRowColInd_R_implementation}
\alias{findDistValueByRowColInd_R_implementation}
\title{Calculate index of the element in distance object based on the row and column index}
\usage{
findDistValueByRowColInd_R_implementation(distance, row_inds, col_inds)
}
\description{
Calculate index of the element in distance object based on the row and column index
}

