% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.clustered}
\alias{is.clustered}
\alias{is.clustered.densityCluster}
\title{Check whether a densityCluster object has been clustered}
\usage{
is.clustered(x)

\method{is.clustered}{densityCluster}(x)
}
\arguments{
\item{x}{the \code{densityCluster} object to check.}
}
\value{
\code{TRUE} if the object has been clustered, otherwise \code{FALSE}.
}
\description{
Check whether a densityCluster object has been clustered
}

