% Generated by roxygen2 (4.0.1): do not edit by hand
\name{localDensity}
\alias{localDensity}
\title{Computes the local density of points in a distance matrix}
\usage{
localDensity(distance, dc, gaussian = FALSE)
}
\arguments{
\item{distance}{a distance matrix.}

\item{dc}{a numeric value specifying the distance cutoff.}

\item{gaussian}{logical indicating that a Gaussian kernel should be used to
  estimate the density.}
}
\value{
A vector of local density values, one for each point.
}
\description{
This function takes a distance matrix and a distance cutoff and calculates
the local density for each point in the matrix. The computation can either be
done using a simple summation of the points within the distance cutoff for
each point, or by applying a Gaussian kernel scaled by the distance cutoff
(more robust for low-density data).
}

