% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.densityCluster}
\alias{plot.densityCluster}
\title{Plot a decision graph for a densityCluster object}
\usage{
\method{plot}{densityCluster}(x, ...)
}
\arguments{
\item{x}{a \code{densityCluster} object as produced by \code{\link{dclust}}.}

\item{...}{extra parameters to pass to \code{\link{plot}}.}
}
\value{
Graphical parameters of the plot (invisibly).
}
\description{
If \eqn{\rho} and \eqn{\delta} thresholds have been specified in the object, points outside
these thresholds will be filled in with color.
}

