% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityClust.R
\name{plotDensClust}
\alias{plotDensClust}
\title{Plot decision graph, gamma graph and MDS scatterplot of observations}
\usage{
plotDensClust(x, dim.x = NULL, dim.y = NULL, num.pts = 20, col = NULL)
}
\arguments{
\item{x}{A densityCluster object as produced by \code{\link{densityClust}}}

\item{dim.x, dim.y}{Vectors giving projections on first and second
multi-dimensonal scaling (principal coordinates) axes. If not given,
MDS is calculated and returned by the function.}

\item{num.pts}{number of points to plot (only used if \code{plot.gamma = TRUE}).}

\item{col}{Vector of colors for clusters.}
}
\description{
This function produces a single figure containing a decision graph
(rho vs. delta), a gamma graph (sorted rho * delta), and a
multi-dimensional scaling (MDS) scatterplot.
}
\examples{
data(iris)
data.dist <- dist(iris[, 1:4])
pca <- princomp(iris[, 1:4])

op <- par(ask = TRUE)
dens.clust <- densityClust(data.dist)
plotDensClust(dens.clust)

n.vec <- 2:3
clust.list <- lapply(n.vec, function(n) {
  new.clust <- findClusters(dens.clust, k = n)
  plotDensClust(new.clust, pca$scores)
  new.clust
})
names(clust.list) <- n.vec
par(op)

}
\seealso{
\code{\link{plot.densityCluster}} \code{\link{plotMDS}}
}

