% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{plotFuncs}
\alias{plotClustMDS}
\alias{plotDecisionGraph}
\alias{plotDensClust}
\alias{plotFuncs}
\alias{plotGammaGraph}
\title{Plotting Functions}
\usage{
plotDecisionGraph(x, col = NULL)

plotGammaGraph(x, n = 20, col = NULL)

plotClustMDS(x, dim.x = NULL, dim.y = NULL, col = NULL)

plotDensClust(x, dim.x = NULL, dim.y = NULL, n = 20, col = NULL)
}
\arguments{
\item{x}{A densityCluster object as produced by \code{\link{densityClust}}}

\item{col}{Vector of colors for clusters.}

\item{n}{Number of observations to plot in Gamma graph.}

\item{dim.x, dim.y}{Vectors giving projections on first and second
multi-dimensonal scaling (principal coordinates) axes. If not given,
MDS is calculated and returned by the function.}
}
\description{
Plotting functions for density cluster algorithm
}
\examples{
data(iris)
data.dist <- dist(iris[, 1:4])
pca <- princomp(iris[, 1:4])

op <- par(ask = TRUE)
dens.clust <- densityClust(data.dist)
plotDensClust(dens.clust)

n.vec <- 2:3
clust.list <- lapply(n.vec, function(n) {
  new.clust <- findClusters(dens.clust, k = n)
  plotDensClust(new.clust, pca$scores)
  new.clust
})
names(clust.list) <- n.vec
par(op)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

