% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityClust.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot observations using multidimensional scaling and colour by cluster}
\usage{
plotMDS(x, dim.x = NULL, dim.y = NULL, col = NULL)
}
\arguments{
\item{x}{A densityCluster object as produced by \code{\link{densityClust}}}

\item{dim.x, dim.y}{Vectors giving projections on first and second
multi-dimensonal scaling (principal coordinates) axes. If not given,
MDS is calculated and returned (invisibly) by the function.}

\item{col}{Vector of colors for clusters.}
}
\description{
This function produces an MDS scatterplot based on the distance matrix of the
densityCluster object, and, if clusters are defined, colours each observation
according to cluster affiliation and labels them. Observations belonging to a
cluster core are plotted with filled circles and observations belonging to
the halo with hollow circles.
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- densityClust(irisDist, gaussian=TRUE)
plot(irisClust) # Inspect clustering attributes to define thresholds

irisClust <- findClusters(irisClust, rho=2, delta=2)
plotMDS(irisClust)
split(iris[,5], irisClust$clusters)

}
\seealso{
\code{\link{densityClust}}
}

