% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMDS}
\alias{plotMDS}
\title{Plot points using multidimensional scaling and color by cluster assignment}
\usage{
plotMDS(x)
}
\arguments{
\item{x}{a \code{densityCluster} object as produced by \code{\link{dclust}}.}
}
\value{
Graphical parameters of the plot (invisibly).
}
\description{
This function produces an MDS scatterplot based on the distance matrix of the
densityCluster object, and, if clusters are defined, colors each observation
according to cluster assignment. Points belonging to a cluster core are
plotted with filled circles and points belonging to the halo with
hollow circles.
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- dclust(irisDist, gaussian=TRUE)
irisClust <- findClusters(irisClust, rho=2, delta=2)
plotMDS(irisClust)
}
\seealso{
\code{\link{dclust}}
}

