% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityClust.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot observations using multidimensional scaling and colour by cluster}
\usage{
plotMDS(x, ...)
}
\arguments{
\item{x}{A densityCluster object as produced by \code{\link[=densityClust]{densityClust()}}}

\item{...}{Additional parameters. Currently ignored}
}
\description{
This function produces an MDS scatterplot based on the distance matrix of the
densityCluster object (if there is only the coordinates information, a distance
matrix will be calculate first), and, if clusters are defined, colours each
observation according to cluster affiliation. Observations belonging to a cluster
core is plotted with filled circles and observations belonging to the halo with
hollow circles. This plotting is not suitable for running large datasets (for example
datasets with > 1000 samples). Users are suggested to use other methods, for example
tSNE, etc. to visualize their clustering results too.
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- densityClust(irisDist, gaussian=TRUE)
plot(irisClust) # Inspect clustering attributes to define thresholds

irisClust <- findClusters(irisClust, rho=2, delta=2)
plotMDS(irisClust)
split(iris[,5], irisClust$clusters)

}
\seealso{
\code{\link[=densityClust]{densityClust()}} for creating \code{densityCluster}
objects, and \code{\link[=plotTSNE]{plotTSNE()}} for an alternative plotting approach.
}
